Passed RadhydroPulseMG test
