+++ src/CMakeLists.txt
@@ -169,8 +169,7 @@ add_subdirectory(RadhydroShockCGS)
 add_subdirectory(RadhydroShockMultigroup)
 add_subdirectory(RadhydroUniformAdvecting)
 add_subdirectory(RadhydroPulse)
-add_subdirectory(RadhydroPulseGrey)
-add_subdirectory(RadhydroPulseMultigroup)
+add_subdirectory(RadhydroPulseMG)
 
 add_subdirectory(ODEIntegration)
 add_subdirectory(Cooling)
