+++ src/RadhydroPulseMG/test_radhydro_pulse_MG.hpp
@@ -1,6 +1,6 @@
-#ifndef TEST_RADHYDRO_PULSE_GREY_HPP_ // NOLINT
-#define TEST_RADHYDRO_PULSE_GREY_HPP_
-/// \file test_radhydro_pulse_grey.hpp
+#ifndef TEST_RADHYDRO_PULSE_MG_HPP_ // NOLINT
+#define TEST_RADHYDRO_PULSE_MG_HPP_
+/// \file test_radiation_marshak.hpp
 /// \brief Defines a test problem for radiation in the static diffusion regime.
 ///
 
@@ -18,4 +18,4 @@
 
 // function definitions
 
-#endif // TEST_RADHYDRO_PULSE_GREY_HPP_
+#endif // TEST_RADHYDRO_PULSE_MG_HPP_
