--- src/RadhydroPulseMultigroup/test_radhydro_pulse_multigroup.hpp
@@ -1,26 +1,0 @@
-#ifndef TEST_RADHYDRO_PULSE_MULTIGROUP_HPP_ // NOLINT
-#define TEST_RADHYDRO_PULSE_MULTIGROUP_HPP_
-//==============================================================================
-// TwoMomentRad - a radiation transport library for patch-based AMR codes
-// Copyright 2020 Benjamin Wibking.
-// Released under the MIT license. See LICENSE file included in the GitHub repo.
-//==============================================================================
-/// \file test_radiation_marshak.hpp
-/// \brief Defines a test problem for radiation in the static diffusion regime.
-///
-
-// external headers
-#ifdef HAVE_PYTHON
-#include "matplotlibcpp.h"
-#endif
-#include <fmt/format.h>
-#include <fstream>
-
-// internal headers
-
-#include "interpolate.hpp"
-#include "radiation_system.hpp"
-
-// function definitions
-
-#endif // TEST_RADHYDRO_PULSE_MULTIGROUP_HPP_
